backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
supabase==2.28.3
openai==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.11
pgvector==0.4.2
langgraph==1.1.6
langchain-openai==1.1.12
langchain-core==1.2.28
celery==5.6.3
redis==7.4.0
httpx
pydantic
